pyproject.toml
[project]
name = "ragapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.0",
    "inngest>=0.5.9",
    "llama-index-llms-gemini>=0.1.7",
    "llama-index-core>=0.14.3",
    "llama-index-readers-file>=0.5.4",
    "openai>=2.0.1",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "streamlit>=1.50.0",
    "uvicorn>=0.37.0",
]
